pyproject.toml
[tool.poetry]
name = "talent-matching"
version = "0.1.0"
description = "Talent Evaluation & Matchmaking System using Dagster"
readme = "README.md"
license = "MIT"
authors = ["Superteam Talent"]
packages = [{include = "talent_matching"}]

[tool.poetry.scripts]
talent-matching = "talent_matching.definitions:main"
remote-ui = "talent_matching.cli:remote_ui"
local-dev = "talent_matching.cli:local_dev"
local-matchmaking = "talent_matching.cli:local_matchmaking"
deploy = "talent_matching.cli:deploy"
setup-db = "talent_matching.cli:setup_db"
with-local-db = "talent_matching.cli:with_local_db"
with-remote-db = "talent_matching.cli:with_remote_db"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"

# Dagster core
dagster = "^1.12.15"
dagster-webserver = "^1.12.15"
dagster-postgres = "^0.28.15"

# Database
psycopg2-binary = ">=2.9.9"
pgvector = ">=0.2.4"
sqlalchemy = ">=2.0.0"
alembic = ">=1.13.0"

# HTTP client (for social API integrations)
httpx = ">=0.26.0"

# Utilities
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"
ruff = ">=0.1.0"
mypy = ">=1.7.0"
pre-commit = ">=3.6.0"

[tool.poetry.group.llm]
optional = true

[tool.poetry.group.llm.dependencies]
openai = ">=1.0.0"
anthropic = ">=0.18.0"
tiktoken = ">=0.5.0"

[tool.poetry.group.parsing]
optional = true

[tool.poetry.group.parsing.dependencies]
pymupdf = ">=1.24.0"
pypdf = ">=3.17.0"
python-docx = ">=1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "talent_matching.definitions"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
